refs CRASM-3004: Separate NavigationContext into context and provider files
- Split NavigationContext.tsx into NavigationContext.ts and NavigationContextProvider.tsx
- Follow established app pattern where contexts and providers are separated
- NavigationContext.ts: Contains interface, context, hook, and helper functions
- NavigationContextProvider.tsx: Contains provider implementation and state management
