+++ frontend/src/context/index.ts
@@ -6,3 +6,5 @@ export * from './SavedSearchContext';
 export * from './SavedSearchContextProvider';
 export * from './FilterDrawerContext';
 export * from './FilterDrawerContextProvider';
+export * from './NavigationContext';
+export * from './NavigationContextProvider';
